+++ src/pretalx/orga/apps.py
@@ -6,5 +6,6 @@ class OrgaConfig(AppConfig):
 
     def ready(self):
         from . import permissions  # noqa
+        from . import receivers  # noqa
         from . import signals  # noqa
         from .phrases import OrgaPhrases  # noqa